backend/backend_fastapi/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pymysql==1.1.1
PyJWT==2.10.1
bcrypt==4.2.1
